# openapi.yaml
openapi: 3.0.3
info:
  title: CSCM Tool API
  description: >
    Self-contained REST API for provisioning and managing Minecraft servers as
    Docker containers, with PlayIT tunnel and Cloudflare DNS automation.
  version: "2.0.0"
servers:
  - url: http://localhost:5000
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Server:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        slug: { type: string }
        type: { type: string, enum: [paper, forge, fabric, vanilla, purpur] }
        version: { type: string }
        port: { type: integer }
        status: { type: string }
        runtime_status:
          type: string
          enum: [not_created, stopped, starting, healthy, unhealthy, running]
        created_at: { type: string }
        tunnels:
          type: array
          items:
            type: object
            properties:
              address: { type: string }
              local_port: { type: integer }
              external_port: { type: integer, nullable: true }
        dns_records:
          type: array
          items:
            type: object
            properties:
              type: { type: string }
              name: { type: string }
              target: { type: string }
              port: { type: integer, nullable: true }
  parameters:
    ServerId:
      name: id
      in: path
      required: true
      schema: { type: integer }
...
paths:
...
  /api/servers:
    get:
      tags: [Servers]
      summary: List all servers
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  servers:
                    type: array
                    items: { $ref: "#/components/schemas/Server" }
    post:
      tags: [Servers]
      summary: Provision a new Minecraft server
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                type:
                  {
                    type: string,
                    enum: [paper, forge, fabric, vanilla, purpur],
                    default: paper,
                  }
                version: { type: string, default: "1.21.4" }
                port:
                  {
                    type: integer,
                    minimum: 1024,
                    maximum: 65535,
                    default: 25565,
                  }
                mem_min: { type: integer, default: 2 }
                mem_max: { type: integer, default: 4 }
                subscription: { type: string, enum: [premium, free] }
                agent: { type: string }
      responses:
        "201": { description: Created }
        "400": { description: Validation error }
        "409": { description: Port conflict }
        "500": { description: Provisioning failed }
  /api/servers/{id}:
    get:
      tags: [Servers]
      summary: Get server detail
      parameters: [{ $ref: "#/components/parameters/ServerId" }]
      responses:
        "200": { description: OK }
        "404": { description: Not found }
    delete:
      tags: [Servers]
      summary: Deprovision a server and all its resources
      parameters: [{ $ref: "#/components/parameters/ServerId" }]
      responses:
        "200": { description: Deleted }
        "404": { description: Not found }
  /api/servers/{id}/start:
    post:
      tags: [Lifecycle]
      parameters: [{ $ref: "#/components/parameters/ServerId" }]
      responses:
        { "200": { description: OK }, "404": { description: Not found } }
  /api/servers/{id}/stop:
    post:
      tags: [Lifecycle]
      parameters: [{ $ref: "#/components/parameters/ServerId" }]
      responses:
        { "200": { description: OK }, "404": { description: Not found } }
  /api/servers/{id}/restart:
    post:
      tags: [Lifecycle]
      parameters: [{ $ref: "#/components/parameters/ServerId" }]
      responses:
        { "200": { description: OK }, "404": { description: Not found } }
  /api/servers/{id}/kill:
    post:
      tags: [Lifecycle]
      parameters: [{ $ref: "#/components/parameters/ServerId" }]
      responses:
        { "200": { description: OK }, "404": { description: Not found } }
  /api/servers/{id}/command:
    post:
      tags: [Console]
      summary: Send an RCON console command
      parameters: [{ $ref: "#/components/parameters/ServerId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [command]
              properties:
                command: { type: string }
      responses:
        "200": { description: OK }
        "400": { description: Missing command }
        "404": { description: Not found }
  /api/servers/{id}/logs:
    get:
      tags: [Console]
      summary: Get recent console log lines
      parameters:
        - { $ref: "#/components/parameters/ServerId" }
        - name: tail
          in: query
          schema: { type: integer, default: 200 }
      responses:
        "200": { description: OK }
...
  /api/servers/{id}/stats:
    get:
      tags: [Console]
      summary: Live resource stats and player list
      parameters: [{ $ref: "#/components/parameters/ServerId" }]
      responses:
        "200": { description: OK }
  /api/servers/{id}/name:
    patch:
      tags: [Modification]
      parameters: [{ $ref: "#/components/parameters/ServerId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties: { name: { type: string } }
      responses:
        { "200": { description: OK }, "404": { description: Not found } }
  /api/servers/{id}/port:
    patch:
      tags: [Modification]
      summary: Change the host port (recreates the container)
      parameters: [{ $ref: "#/components/parameters/ServerId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [port]
              properties:
                { port: { type: integer, minimum: 1024, maximum: 65535 } }
      responses:
        "200": { description: OK }
        "404": { description: Not found }
        "409": { description: Port conflict }
  /api/servers/{id}/ram:
    patch:
      tags: [Modification]
      summary: Change JVM heap allocation (recreates the container)
      parameters: [{ $ref: "#/components/parameters/ServerId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mem_min, mem_max]
              properties:
                mem_min: { type: integer }
                mem_max: { type: integer }
      responses:
        { "200": { description: OK }, "404": { description: Not found } }
...
  /api/servers/{id}/tunnel:
    post:
      tags: [Networking]
      parameters: [{ $ref: "#/components/parameters/ServerId" }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                region: { type: string }
                subscription: { type: string, enum: [premium, free] }
                agent: { type: string }
      responses:
        { "201": { description: Created }, "404": { description: Not found } }
  /api/servers/{id}/subdomain:
    patch:
      tags: [Networking]
      parameters: [{ $ref: "#/components/parameters/ServerId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [subdomain]
              properties: { subdomain: { type: string } }
      responses:
        { "200": { description: OK }, "404": { description: Not found } }